updated project-completely new methods of doing things

+++ src/org/marcus/weather/HourlyForecast.java
@@ -3,10 +3,10 @@ package org.marcus.weather;
 public class HourlyForecast {
 
 	public int hour;
-	public Float temp;
+	public Integer temp;
 	public Integer PoP;
 	
-	public HourlyForecast(int time, Float htemp, Integer hPoP){
+	public HourlyForecast(int time, Integer htemp, Integer hPoP){
 		hour = time;
 		temp = htemp;
 		PoP = hPoP;
